fix(#601): respond to PR review — extract chart formatters + YTD clamp
WARNING: formatHoverLabel / tickFormatter / humanizeVolume were
duplicated verbatim in PriceChart and ChartWorkspaceCanvas. Bot
flagged that a future patch to one would silently diverge the
overview pane from the workspace. Extracted to
frontend/src/lib/chartFormatters.ts; both consumers now import the
shared symbols.

NITPICK: YTD on Jan 1 returned days=0 → empty state. Clamped to
max(1, days) so the chart shows yesterday's bar at minimum.

Frontend gates: typecheck clean; 568/568 tests pass.
Backend gates: ruff clean; candles tests 6/6.
